Removed the node when creating a new board
Instead of removing the nodes I was just removing the background-color property, but the nodes still existed and when creating a new board the alpha value of rgba would be saved.  Also added a conditional so the alpha doesn't go over 1, it worked without it but this is probably better.